move interval to solver interface
